Write a program to change an infix expression to postfix expresssion
My computer, Mac, really hates every time I write an infix expression! You have to ease the task for Mac by transforming a given Infix expression to Postfix.

Instead of numeric variables the expression given is using lowercase english letters (a to z). The priority of operators is as follows '+', '-', '*', '/', '^' (increasing left to right) '^' is for exponentiation.

Each case would be a string without space in valid form. There won't be any case with ambiguity i.e., such that two different answers are possible e.g., (a+b+c), it would be well parenthesized in that case.